Left,Rightそれぞれのコンポーネント作成し、4プレイヤー分ブロック作成
