+++ src/App.vue
@@ -1,5 +1,17 @@
 <template>
   <div id="app">
+    <LeftPlayersBlock 
+    class="player1"
+    @click="handClick3"
+    >
+    </LeftPlayersBlock>
+
+    <LeftPlayersBlock 
+    class="player2"
+    @click="handClick3"
+    >
+    </LeftPlayersBlock>
+
     <!-- ブロックを配置する基盤 -->
     <div class="table">
       <table>
@@ -8,7 +20,7 @@
           :key="y"
         >
           <td 
-            class="MainTableCell"
+            class="MainTableCell" 
             group="items" 
             v-for="(cell, x) in boardRow" 
             :key="x" 
@@ -21,330 +33,34 @@
         </tr>
       </table>
     </div>
-    <!--// 基盤 -->
-    <div class="player1">
-    <!-- 手持ちのブロック -->
-      <div class="handBlock">
-        <div 
-          v-for="(block, bi) in blocks" 
-          :key="bi" 
-          group="items" 
-          @click="handClick(block)"
-        >
-          <table>
-            <tr v-for="(blockTR, yi) in block.data" :key="yi">
-              <td 
-                class="blockCell" 
-                v-for="(blockTD, xi) in blockTR" 
-                :key="xi" 
-                :class="`is-${blockTD.code}`"
-              >
-              </td>
-            </tr>
-          </table>
-        </div>
-      </div>
-      <!-- // 手持ちのブロック -->
-      <!-- 選んだブロックの表示 -->
-      <div class="chooseblocks">
-        <draggable tag="table">
-          <tr 
-            v-for="(chooseBlockRow, vi) in chooseBlock"
-            :key="vi">
-            <td 
-              class="blockCell2"
-              v-for="(chooseBlockCell, mi) in chooseBlockRow" 
-              :key="mi"
-              :class="`is-${chooseBlockCell.code}`">
-            </td>
-          </tr>
-        </draggable>
-        <button @click.prevent @click="turnLeft()">左に回転
-        </button>
-        <button @click.prevent @click="turnRight()">右に回転
-        </button>
-        <button @click.prevent @click="Inversion()">反転
-        </button>
-      </div>
-    <!--// 選んだブロックの表示 -->
-    </div>
-    <div class="player2">
-    <!-- 手持ちのブロック -->
-      <div class="handBlock">
-        <div 
-          v-for="(block, bi) in blocks" 
-          :key="bi" 
-          group="items" 
-          @click="handClick(block)"
-        >
-          <table>
-            <tr v-for="(blockTR, yi) in block.data" :key="yi">
-              <td 
-                class="blockCell" 
-                v-for="(blockTD, xi) in blockTR" 
-                :key="xi" 
-                :class="`is-${blockTD.code}`"
-              >
-              </td>
-            </tr>
-          </table>
-        </div>
-      </div>
-      <!-- // 手持ちのブロック -->
-      <!-- 選んだブロックの表示 -->
-      <div class="chooseblocks">
-        <draggable tag="table">
-          <tr 
-            v-for="(chooseBlockRow, vi) in chooseBlock"
-            :key="vi">
-            <td 
-              class="blockCell2"
-              v-for="(chooseBlockCell, mi) in chooseBlockRow" 
-              :key="mi"
-              :class="`is-${chooseBlockCell.code}`">
-            </td>
-          </tr>
-        </draggable>
-        <button @click.prevent @click="turnLeft()">左に回転
-        </button>
-        <button @click.prevent @click="turnRight()">右に回転
-        </button>
-        <button @click.prevent @click="Inversion()">反転
-        </button>
-      </div>
-    <!--// 選んだブロックの表示 -->
-    </div>
-    <div class="player3">
-    <!-- 手持ちのブロック -->
-      <div class="handBlock">
-        <div 
-          v-for="(block, bi) in blocks" 
-          :key="bi" 
-          group="items" 
-          @click="handClick(block)"
-        >
-          <table>
-            <tr v-for="(blockTR, yi) in block.data" :key="yi">
-              <td 
-                class="blockCell" 
-                v-for="(blockTD, xi) in blockTR" 
-                :key="xi" 
-                :class="`is-${blockTD.code}`"
-              >
-              </td>
-            </tr>
-          </table>
-        </div>
-      </div>
-      <!-- // 手持ちのブロック -->
-      <!-- 選んだブロックの表示 -->
-      <div class="chooseblocks">
-        <draggable tag="table">
-          <tr 
-            v-for="(chooseBlockRow, vi) in chooseBlock"
-            :key="vi">
-            <td 
-              class="blockCell2"
-              v-for="(chooseBlockCell, mi) in chooseBlockRow" 
-              :key="mi"
-              :class="`is-${chooseBlockCell.code}`">
-            </td>
-          </tr>
-        </draggable>
-        <button @click.prevent @click="turnLeft()">左に回転
-        </button>
-        <button @click.prevent @click="turnRight()">右に回転
-        </button>
-        <button @click.prevent @click="Inversion()">反転
-        </button>
-      </div>
-    <!--// 選んだブロックの表示 -->
-    </div>
-    <PlayersBlock class="player4"></PlayersBlock>
+    <!-- //ブロックを配置する基盤 -->
+
+    <RightPlayersBlock 
+    class="player3"
+    @click="handClick3"
+    >
+    </RightPlayersBlock>
+    
+    <RightPlayersBlock 
+    class="player4"
+    @click="handClick4"
+    >
+    </RightPlayersBlock>
   </div>
 </template>
 
 <script>
 
-import draggable from 'vuedraggable'
-import PlayersBlock from './components/PlayersBlock.vue'
+import RightPlayersBlock from './components/RightPlayersBlock.vue'
+import LeftPlayersBlock from './components/LeftPlayersBlock.vue'
+// import "./style.scss";
 export default {
   components: {
-    draggable,
-    PlayersBlock,
+    RightPlayersBlock,
+    LeftPlayersBlock,
   },
   name: 'App',
   data (){
-    const blocks = [
-      { name:0,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:1,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:2,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:3,
-        data:[
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:4,
-        data:[
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:5,
-        data:[
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:6,
-        data:[
-          [{code:1}, {code:1}, {code:1}, {code:1}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:7,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:8,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:9,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:1}, {code:1}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:10,
-        data:[
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:11,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:12,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:13,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:14,
-        data:[
-          [{code:1}, {code:1}, {code:1}, {code:1}, {code:1}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:15,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:16,
-        data:[
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:17,
-        data:[
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:18,
-        data:[
-          [{code:0}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:19,
-        data:[
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:1}, {code:1}, {code:1}, {code:0}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-      { name:20,
-        data:[
-          [{code:1}, {code:1}, {code:1}, {code:1}, {code:0}],
-          [{code:0}, {code:1}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-          [{code:0}, {code:0}, {code:0}, {code:0}, {code:0}],
-      ]},
-    ];
     const mainBoard = [];
       for (let y = 0; y < 20; y++ ) {
         const boardRow = []
@@ -358,71 +74,31 @@ export default {
         }
         mainBoard.push(boardRow)
       } 
-    const chooseBlock = [];
-      for (let y = 0; y < 5; y++ ) {
-        const chooseBlockRow = []
-        for (let x = 0; x < 5; x++){
-          const chooseBlockCell = {
-            y: y,
-            x: x,
-            code: 0,
-          };
-        chooseBlockRow.push(chooseBlockCell)
-      }
-      chooseBlock.push(chooseBlockRow)
-      }
 // ここからvue.js記述
     return{
       mainBoard: mainBoard,
-      blocks:blocks,
-      dragItem: "",
-      item: [],
-      haveItem: [],
-      chooseBlock: chooseBlock,
+      item1: [],
+      item2: [],
+      item3: [],
+      item4: [],
     };
   },
   methods:{
-    turnLeft(){
-      this.haveItem = JSON.parse(JSON.stringify(this.chooseBlock))
-      for (let i = 0; i < 5; i++) {
-        for (let j = 0; j < 5; j++) {
-          this.chooseBlock[4 - i][j].code = this.haveItem[j][i].code
-        }
-      }
+    handClick3(chooseBlock){
+      this.item3 = chooseBlock
     },
-    turnRight(){
-      this.haveItem = JSON.parse(JSON.stringify(this.chooseBlock))
-      for (let i = 0; i < 5; i++) {
-        for (let j = 0; j < 5; j++) {
-          this.chooseBlock[i][4 - j].code = this.haveItem[j][i].code
-        }
-      }
-    },
-    Inversion(){
-      this.haveItem = JSON.parse(JSON.stringify(this.chooseBlock))
-      for (let i = 0; i < 5; i++) {
-        for (let j = 0; j < 5; j++) {
-          this.chooseBlock[i][j].code = this.haveItem[j][i].code
-        }
-      }
-    },
-    handClick(block){
-      this.dragItem = block.name;
-      this.item = this.blocks.find(x=>x.name==block.name).data;
-      for (let i = 0; i < 5; i++) {
-        for (let j = 0; j < 5; j++) {
-          this.chooseBlock[i][j].code = this.item[i][j].code
-        }
-      }
+    handClick4(chooseBlock){
+      this.item4 = chooseBlock
     },
     droped(x, y){
-      for (let i = 0; i < 5; i++) {
-        for (let j = 0; j < 5; j++) {
-          if (this.item[i][j].code==1) {
-            this.mainBoard[y + i][x + j].code = this.item[i][j].code
+      if (this.item3.length == 5) {
+        for (let i = 0; i < 5; i++) {
+          for (let j = 0; j < 5; j++) {
+            this.mainBoard[y + i][x + j].code = this.item3[i][j].code;
           }
         }
       }
+      this.item3 = []
     },
   }
 }
@@ -468,27 +144,27 @@ export default {
   display: flex;
 }
 
+
 .MainTableCell{
-    height: 20px;
-    width: 20px;
-    border: solid black 1px;
+  height: 20px;
+  width: 20px;
+  border: solid black 1px;
 }
 
 .blockCell{
-    height: 6px;
-    width: 6px;
+  height: 6px;
+  width: 6px;
 }
 .blockCell2{
-    height: 16px;
-    width: 16px;
-
+  height: 16px;
+  width: 16px;
 }
 
 .is-2{
-    background-color: blueviolet;
+  background-color: blueviolet;
 }
 .is-1{
-    background-color: rgb(43, 144, 226);
+  background-color: rgb(43, 144, 226);
 }
 
 .handBlock{
